requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
babel==2.6.0              # via django-phonenumber-field
django-bootstrap4==0.0.6
django-markupfield==1.5.0
django-phonenumber-field==2.0.0
django==2.1
docutils==0.14
markdown==2.6.11
phonenumberslite==8.9.10  # via django-phonenumber-field
pytz==2018.5              # via babel, django
